+++ tests/perf/test-call-bound-deep.js
@@ -13,6 +13,9 @@ function test() {
     bound = bound.bind('dummy-this', 9);
     bound = bound.bind('dummy-this', 10);
     var i;
+
+    var t1 = Date.now();
+
     for (i = 0; i < 1e5; i++) {
         void bound(11, 12); void bound(11, 12); void bound(11, 12); void bound(11, 12); void bound(11, 12);
         void bound(11, 12); void bound(11, 12); void bound(11, 12); void bound(11, 12); void bound(11, 12);
@@ -35,7 +38,11 @@ function test() {
         void bound(11, 12); void bound(11, 12); void bound(11, 12); void bound(11, 12); void bound(11, 12);
         void bound(11, 12); void bound(11, 12); void bound(11, 12); void bound(11, 12); void bound(11, 12);
     }
+
+    var t2 = Date.now();
+    print((1e5 * 100 / (t2 - t1)) + ' calls per millisecond');
 }
+
 try {
     test();
 } catch (e) {